+++ if_possible_to_fill_figure.c
@@ -3,7 +3,7 @@
 #include "fillit.h"
 #define EMPTY '0'
 
-int	is_empty_and_correct(t_term *figure, size_t size, char **map)
+int	is_empty_and_correct(t_term *figure, int size, char **map)
 {
   int i;
 
@@ -17,7 +17,7 @@ int	is_empty_and_correct(t_term *figure, size_t size, char **map)
   return (1);
 }
 
-int	if_possible(char **map, size_t size, t_term *figure, int i, int j)
+int	if_possible(char **map, int size, t_term *figure, int i, int j)
 {
   int change_i;
   int change_j;
